Support leading/trailing sign bits in fixedpoint.  Closes MYL-72.
These updates allow specification of Q0.000015 numbers (in 16 bit word length)
via a positive point_shift dtype value.  The magnitude specifies the number of
leading zeros (or 1's if negative) to assume is present in internal
representation.  Similarly, we support additional trailing zeros to the integer
portion of the number with a negative point_shift value.